+++ portal/src/util/resource.ts
